backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
httpx==0.27.2
tenacity==9.0.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
celery[redis]==5.4.0
structlog==24.4.0
Pillow==10.4.0
psutil==6.1.0
sentry-sdk[fastapi]==2.19.0
aiofiles==24.1.0
python-levenshtein==0.26.0
uvloop==0.21.0
email-validator==2.1.0.post1
slowapi==0.1.9